Ensure fail2ban actually installs firewalld hooks
We try to run things in the right order, *and* we check if fail2ban's
ipset rule is known to firewalld. (We restart fail2ban if not, which
will recreate the rule.)

The task that checks whether the rule exists is always marked as a
change -- I'd like to figure out how to avoid that.